telnet added, full ota, feedback nrf24

+++ src/varaibles.cpp
@@ -3,7 +3,7 @@
 MessageData currentBoatInfo;
 bool boatInfoTransmitter;
 uint16_t lastTransmitterSendTime = 0;
-uint16_t last_message_milis = 0;
+unsigned long last_message_milis = 0;
 bool IS_STOPPED;
 
 uint8_t LAST_UPDATE_DURATION;
